requirements.txt
mysql-connector-python>=8.0.33
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.2.0
statsmodels>=0.14.0
scipy>=1.10.0
ta>=0.10.2
plotly>=5.14.0
requests>=2.28.0
python-dateutil>=2.8.2